+++ Assets/_Scripts/misc/Matrix Helper.cs
@@ -57,19 +57,19 @@ public static class MatrixHelper
         return newMat;
     }
     
-    public static Vector2Int ConvertToIndex(Vector2Int boardPos, int boardSize)
+    public static Vector2Int ConvertToIndex(Vector2Int boardPos, uint boardSize)
     {
         Vector2Int index = new Vector2Int();
-        index.x = boardSize - boardPos.y;
+        index.x = (int)boardSize - boardPos.y;
         index.y = boardPos.x;
 
         return index;
     }
     
-    public static Vector2Int ConvertToPosition(Vector2Int index, int boardSize)
+    public static Vector2Int ConvertToPosition(Vector2Int index, uint boardSize)
     {
         Vector2Int position = new Vector2Int();
-        position.x = boardSize - index.y;
+        position.x = (int)boardSize - index.y;
         position.y = index.x;
 
         return index;
